refactor: split missing tests into modules
