+++ apollo-federation/tests/query_plan/build_query_plan_tests/abstract_types.rs
@@ -1,296 +1,3 @@
-#[test]
-fn composite_field_with_subfields_from_another_service() {
-    let planner = planner!(
-        accounts: r#"
-        type Query {
-            me: User
-        }
-
-        type User @key(fields: "id") {
-            id: ID!
-            name: Name
-            birthDate: String
-        }
-
-        type Name {
-            first: String
-        }
-        "#,
-        reviews: r#"
-        type Query {
-            topReviews: [Review!]!
-        }
-
-        type Review {
-            body: String!
-            author: User
-        }
-
-        type User @key(fields: "id") {
-            id: ID!
-        }
-        "#,
-    );
-    assert_plan!(
-        &planner,
-        r#"
-          query {
-            topReviews {
-              body
-              author {
-                name { first }
-              }
-            }
-          }
-        "#,
-        @r###"
-        QueryPlan {
-          Sequence {
-            Fetch(service: "reviews") {
-              {
-                topReviews {
-                  body
-                  author {
-                    __typename
-                    id
-                  }
-                }
-              }
-            },
-            Flatten(path: "topReviews.@.author") {
-              Fetch(service: "accounts") {
-                {
-                  ... on User {
-                    __typename
-                    id
-                  }
-                } =>
-                {
-                  ... on User {
-                    name {
-                      first
-                    }
-                  }
-                }
-              },
-            },
-          },
-        }
-        "###
-    );
-}
-
-#[test]
-fn composite_field_parent_selection_empty() {
-    let planner = planner!(
-        accounts: r#"
-        type Query {
-            me: User
-        }
-
-        type User @key(fields: "id") {
-            id: ID!
-            name: Name
-            birthDate: String
-        }
-
-        type Name {
-            first: String
-        }
-        "#,
-        reviews: r#"
-        type Query {
-            topReviews: [Review!]!
-        }
-
-        type Review {
-            author: User
-        }
-
-        type User @key(fields: "id") {
-            id: ID!
-        }
-        "#,
-    );
-    assert_plan!(
-        &planner,
-        r#"
-          query {
-            topReviews {
-              author {
-                name { first }
-              }
-            }
-          }
-        "#,
-        @r###"
-        QueryPlan {
-          Sequence {
-            Fetch(service: "reviews") {
-              {
-                topReviews {
-                  author {
-                    __typename
-                    id
-                  }
-                }
-              }
-            },
-            Flatten(path: "topReviews.@.author") {
-              Fetch(service: "accounts") {
-                {
-                  ... on User {
-                    __typename
-                    id
-                  }
-                } =>
-                {
-                  ... on User {
-                    name {
-                      first
-                    }
-                  }
-                }
-              },
-            },
-          },
-        }
-        "###
-    );
-}
-
-#[test]
-fn field_requires_value_from_base_service() {
-    let planner = planner!(
-        product: r#"
-        type Query {
-            topCars: [Car!]
-        }
-
-        type Car @key(fields: "id") {
-            id: ID!
-            price: Int
-        }
-        "#,
-        reviews: r#"
-        type Car @key(fields: "id") {
-            id: ID!
-            price: Int @external
-            retailPrice: Int @requires(fields: "price")
-        }
-        "#,
-    );
-    assert_plan!(
-        &planner,
-        r#"
-          query {
-            topCars {
-              retailPrice
-            }
-          }
-        "#,
-        @r###"
-        QueryPlan {
-          Sequence {
-            Fetch(service: "product") {
-              {
-                topCars {
-                  __typename
-                  id
-                  price
-                }
-              }
-            },
-            Flatten(path: "topCars.@") {
-              Fetch(service: "reviews") {
-                {
-                  ... on Car {
-                    __typename
-                    id
-                    price
-                  }
-                } =>
-                {
-                  ... on Car {
-                    retailPrice
-                  }
-                }
-              },
-            },
-          },
-        }
-        "###
-    );
-}
-
-#[test]
-fn relationship_field_with_extension_subfields() {
-    let planner = planner!(
-        accounts: r#"
-        type User @key(fields: "id") {
-            id: ID!
-            birthDate: String
-        }
-        "#,
-        reviews: r#"
-        type Query {
-            topReviews: [Review]
-        }
-
-        type Review {
-            author: User
-        }
-
-        type User @key(fields: "id") {
-            id: ID!
-        }
-        "#,
-    );
-    assert_plan!(
-        &planner,
-        r#"
-          query {
-            topReviews {
-              author {
-                birthDate
-              }
-            }
-          }
-        "#,
-        @r###"
-        QueryPlan {
-          Sequence {
-            Fetch(service: "reviews") {
-              {
-                topReviews {
-                  author {
-                    __typename
-                    id
-                  }
-                }
-              }
-            },
-            Flatten(path: "topReviews.@.author") {
-              Fetch(service: "accounts") {
-                {
-                  ... on User {
-                    __typename
-                    id
-                  }
-                } =>
-                {
-                  ... on User {
-                    birthDate
-                  }
-                }
-              },
-            },
-          },
-        }
-        "###
-    );
-}
-
 // Abstract type related tests
 #[test]
 fn adds_typename_when_fetching_interface_fields() {
@@ -920,196 +627,3 @@ fn nested_unions_with_inline_fragments() {
         "###
     );
 }
-
-#[test]
-fn inline_fragment_duplication_deduplicated() {
-    let planner = planner!(
-        documents: r#"
-        type Query {
-            body: Body
-        }
-
-        union Body = Image | Text
-
-        type Image {
-            attributes: TextAttributes
-        }
-
-        type Text {
-            attributes: TextAttributes
-        }
-
-        type TextAttributes {
-            bold: Boolean
-            text: String
-        }
-        "#,
-    );
-    assert_plan!(
-        &planner,
-        r#"
-          query {
-            body {
-              ... on Body {
-                ... on Text {
-                  attributes {
-                    bold
-                    text
-                  }
-                }
-              }
-              ... on Text {
-                attributes {
-                  bold
-                  text
-                }
-              }
-            }
-          }
-        "#,
-        @r###"
-        QueryPlan {
-          Fetch(service: "documents") {
-            {
-              body {
-                __typename
-                ... on Text {
-                  attributes {
-                    bold
-                    text
-                  }
-                }
-              }
-            }
-          },
-        }
-        "###
-    );
-}
-
-#[test]
-fn named_fragment_spread_duplication_deduplicated() {
-    let planner = planner!(
-        documents: r#"
-        type Query {
-            body: Body
-        }
-
-        union Body = Image | Text
-
-        type Image {
-            attributes: TextAttributes
-        }
-
-        type Text {
-            attributes: TextAttributes
-        }
-
-        type TextAttributes {
-            bold: Boolean
-            text: String
-        }
-        "#,
-    );
-    assert_plan!(
-        &planner,
-        r#"
-          fragment TextFragment on Text {
-            attributes {
-              bold
-              text
-            }
-          }
-
-          query {
-            body {
-              ... on Body {
-                ...TextFragment
-              }
-              ...TextFragment
-            }
-          }
-        "#,
-        @r###"
-        QueryPlan {
-          Fetch(service: "documents") {
-            {
-              body {
-                __typename
-                ... on Text {
-                  attributes {
-                    bold
-                    text
-                  }
-                }
-              }
-            }
-          },
-        }
-        "###
-    );
-}
-
-#[test]
-fn overridden_field_and_type() {
-    let planner = planner!(
-        books: r#"
-        type Query {
-            library(id: ID!): Library
-        }
-
-        type Library @key(fields: "id") {
-            id: ID!
-            name: String
-        }
-        "#,
-        accounts: r#"
-        type Library @key(fields: "id") {
-            id: ID!
-            description: String
-        }
-        "#,
-    );
-    assert_plan!(
-        &planner,
-        r#"
-          query {
-            library(id: "3") {
-              name
-              description
-            }
-          }
-        "#,
-        @r###"
-        QueryPlan {
-          Sequence {
-            Fetch(service: "books") {
-              {
-                library(id: 3) {
-                  __typename
-                  id
-                  name
-                }
-              }
-            },
-            Flatten(path: "library") {
-              Fetch(service: "accounts") {
-                {
-                  ... on Library {
-                    __typename
-                    id
-                  }
-                } =>
-                {
-                  ... on Library {
-                    description
-                  }
-                }
-              },
-            },
-          },
-        }
-        "###
-    );
-}
-
